Fix some columns to pl.UInt32 so dataframes have same schema both with "polars" and "polars-u64-idx".
Fix some columns to pl.UInt32 so dataframes have same schema both
with "polars" and "polars-u64-idx".

+++ src/pycisTopic/lda_models.py
@@ -634,7 +634,7 @@ class LDAMallet(utils.SaveLoad, basemodel.BaseTopicModel):
             )
             .lazy()
             .group_by(["topic", "region"])
-            .agg(pl.len().alias("occurrence"))
+            .agg(pl.len().cast(pl.UInt32).alias("occurrence"))
             .collect()
         )
